Relationship mapping improvements
