+++ api/src/main/java/org/openmrs/module/relationshipmapping/api/db/MasterCardRecordDao.java
@@ -1,15 +1,14 @@
 package org.openmrs.module.relationshipmapping.api.db;
 
+import java.util.List;
+
 import org.openmrs.module.relationshipmapping.model.MasterCardObsGroup;
 import org.openmrs.module.relationshipmapping.model.MasterCardObsGroupMember;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Date;
-import java.util.List;
-
 public interface MasterCardRecordDao {
     @Transactional
-    List<MasterCardObsGroup> getNewMasterCardObsGroups(final Date previousExecutionDate);
+    List<MasterCardObsGroup> getNewMasterCardObsGroups();
 
     @Transactional
     List<MasterCardObsGroupMember> getMasterCardObsGroupMembers(final int obsGroupId);
